MASSIVE bug fixes and test updates. (Rebased 61 commits).
- Fixed a bug when variants overlap the end of transcripts.

- Updated logging in FuncotatorUtils::getAlignedRefAllele.

- Fixed a bug in identifying splice sites for intronic regions.

- Fixed a bug in FuncotatorUtils::isIndelBetweenCodons that caused issues
on reverse stranded variants.

- Added in regression test data input files and expected files.
- Updated regression test sets to include only unique variants.
- Added in a 5' flank and a Start Codon insertion to regression test set.

- Fixed a bug in FuncotatorUtils::isIndelBetweenCodons

- Finally fixed a bug with indels and start codons:
Now indels in start codons will not have protein renderings, nor will
they have codon change strings.  This brings Funcotator closer to
Oncotator functionality (in Oncotator, start codon insertions/deletions
    do not have protein change strings or codon change strings).

- Fixed a bug in ordering transcripts by appris ranking.

- Fixed a minor bug in how other transcripts are generated:.
With RNA/LINCRNA transcripts, the protein change would be null and was
append ed to the end of each `other transcript`.  Now the null is no
longer appended.

- Found a bug in insertions on the - strand:
All insertions on - strand have incorrectly rendered genome changes,
    coding sequence changes, and protein changes.
    This is due to how these fields are generated from the reference
    sequence.

- Fixed a bug (insertions on - strand):
Insertions on the - strand would have incorrect reference
sequences/alleles.
Now they are handled as a special case when computing the aligned
reference allele.

- Fixed a bug in transcript selection for GencodeFuncotationFactory:
The LocusLevel / Curation Level was being incorrectly pulled from the
GENE features, rather than the TRANSCRIPT features that contain each
variant.  As a result, the order in which representative transcripts
were chosen was wrong.  The TRANSCRIPT feature is now being used to
determine the Locus/Curation Level.

- `TranscriptType` now determined by transcript annotation, not gene annotation
- Start/stop codon overlapping now corrected for preceding indel bases (is now correct for more cases).
- Changed algorithm for how 5'UTRs are determined.
- Added missing test stubs (must be filled in).

- Refactored how frameshift indels have codon change strings created.

- Added in helper some scripts for testing funcotator.

- Fixed how codon change strings are rendered to be consistent and more
correct.

- Fixed Protein Change strings to be consistent and more
correct.

- Implemented ProteinChangeString for In-Frame, codon-aligned deletions
and insertions (insertions that occur between 2 codons of length n*3, and
    deletions of full codons).
- Implemented ProteinChangeString for In-Frame deletions within codons
(deletions of length n*3 that begin with a base that is not the first
 base of a codon).

- Implemented tests for CreateProteinChangeInfo
- Implemented tests for RenderProteinChangeString
- Implemented tests for IsIndelBetweenCodons
- Implemented tests for GetCodonChangeString

- Added a unit test for
testCreateGencodeFuncotationBuilderWithTrivialFieldsPopulated.

- Fixed a bug when variant ref allele doesn't match reference genome.

- Fixed test cases for - strand indel cdna strings:
There is a bug in oncotator that was fixed in Funcotator involving cdna
strings for - strand indels.  In Oncotator the positions reported are off by 1 (they
    should be one less) and the base reported is also wrong.
This is now fixed.

- Need to fix the implementataion of the manual annotations test.
Right now it's commented out.

- Removed some old code that had been taken out of the main codepath.

- Fixed a bug in how the gencode reference contexts are created.
- Fixed a bug in how the end points for the gencode annotations are
created.

- Ref context field is now consistent for indels.
The reference context will give WINDOW bases before and after the
logical reference allele for a variant.  This is NOT the allele in the
input VCF, but rather the allele that actually has changed.  For
insertions, the logical allele is the SPACE BETWEEN TWO BASES (and
    therefore the resulting string will always be 2xWINDOW bases long).
For deletions, the logical allele is the given ref allele without the
required preceding base.  For MNPs the logical allele is the given ref
allele.
Updated some tests and test data to reflect this change.

- Added a small HG38 regression test set.

- Fixed a boundary bug with codon strings.
Now codon change strings have an alternate (correct) form for insertions
that involve the start codon on the - strand, and the stop codon on the
+ strand.  This form eliminates any overrun/out of bounds exceptions.

- Fixed an issue involving variants that overrun the end of the coding sequence.